redid nav, removed react navigation
