+++ FrugalApp/app/(tabs)/heart.tsx
@@ -1,0 +1,10 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+export default function heart() {
+  return (
+    <View>
+      <Text>heart</Text>
+    </View>
+  )
+}
