+++ FrugalApp/app/index.tsx
@@ -1,0 +1,5 @@
+import { Redirect } from "expo-router";
+
+export default function StartPage() {
+  return <Redirect href="/home" />;
+}
